WEBAPP-223: Refactor fetchers and rename to endpoints
