+++ src/modules/endpoint/endpoints/languages.js
@@ -2,19 +2,16 @@
 
 import LanguageModel from '../models/LanguageModel'
 import { apiUrl } from '../constants'
-import type { Dispatch } from 'redux-first-router/dist/flow-types'
 import EndpointBuilder from '../EndpointBuilder'
-import Payload from '../Payload'
 
 type Params = {
   city: string
 }
 
-export default (dispatch: Dispatch, oldPayload: Payload, params: Params): Promise<Payload> => new EndpointBuilder('languages')
+export default new EndpointBuilder('languages')
   .withParamsToUrlMapper((params: Params): string => `${apiUrl}/${params.city}/de/wp-json/extensions/v0/languages/wpml`)
   .withMapper((json: any): Array<LanguageModel> => json
     .map(language => new LanguageModel(language.code, language.native_name))
     .sort((lang1, lang2) => lang1.code.localeCompare(lang2.code))
   )
   .build()
-  .fetchData(dispatch, oldPayload, params)
