+++ src/modules/endpoint/reducers/index.js
@@ -1,63 +1,26 @@
-// @flow
+import languagesEndpoint from '../endpoints/languages'
+import citiesEndpoint from '../endpoints/cities'
+import categoriesEndpoint from '../endpoints/categories'
+import eventsEndpoint from '../endpoints/events'
+import disclaimerEndpoint from '../endpoints/disclaimer'
+import extrasEndpoint from '../endpoints/extras'
+import sprungbrettJobEndpoint from '../endpoints/sprungbrettJobs'
 
-import { handleAction, handleActions } from 'redux-actions'
-import {
-  CITIES_FETCHED, LANGUAGES_FETCHED, EXTRAS_FETCHED, EVENTS_FETCHED, DISCLAIMER_FETCHED, SPRUNGBRETT_JOBS_FETCHED,
-  CATEGORIES_FETCHED
-} from '../actions/fetcher'
-
-import {
-  CATEGORIES_REMOVED, DISCLAIMER_REMOVED, EVENTS_REMOVED, EXTRAS_REMOVED, LANGUAGES_REMOVED,
-  SPRUNGBRETT_JOBS_REMOVED
-} from '../actions/remover'
-
-import type { Action, Store } from 'redux-first-router/dist/flow-types'
-import Payload from '../Payload'
-
-const fetcherReducer = (state: Store, action: Action) => action.payload
-const removerReducer = () => null
-
-const defaultState = new Payload(false)
-
-// reducers to handle the fetching and removing of data to and from the redux store
-const reducers = {
-  cities: handleAction(CITIES_FETCHED, fetcherReducer, defaultState),
-  languages: handleActions({
-    [LANGUAGES_FETCHED]: fetcherReducer,
-    [LANGUAGES_REMOVED]: removerReducer
-  },
-  defaultState
-  ),
-  categories: handleActions({
-    [CATEGORIES_FETCHED]: fetcherReducer,
-    [CATEGORIES_REMOVED]: removerReducer
-  },
-  defaultState
-  ),
-  events: handleActions({
-    [EVENTS_FETCHED]: fetcherReducer,
-    [EVENTS_REMOVED]: removerReducer
-  },
-  defaultState
-  ),
-  extras: handleActions({
-    [EXTRAS_FETCHED]: fetcherReducer,
-    [EXTRAS_REMOVED]: removerReducer
-  },
-  defaultState
-  ),
-  disclaimer: handleActions({
-    [DISCLAIMER_FETCHED]: fetcherReducer,
-    [DISCLAIMER_REMOVED]: removerReducer
-  },
-  defaultState
-  ),
-  sprungbrettJobs: handleActions({
-    [SPRUNGBRETT_JOBS_FETCHED]: fetcherReducer,
-    [SPRUNGBRETT_JOBS_REMOVED]: removerReducer
-  },
-  defaultState
-  )
-}
+/**
+ * Contains all reducers from all endpoints which are defined in {@link './endpoints/'}
+ */
+const endpoints = [
+  languagesEndpoint,
+  citiesEndpoint(),
+  categoriesEndpoint,
+  disclaimerEndpoint,
+  eventsEndpoint,
+  extrasEndpoint,
+  sprungbrettJobEndpoint
+]
+const reducers = endpoints.reduce((result, endpoint) => {
+  result[endpoint.stateName] = endpoint.createReducer()
+  return result
+}, {})
 
 export default reducers
